amr_deep_learning/requirements.txt
absl-py==0.13.0
aiobotocore==2.13.0
aiofiles==23.2.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
beautifulsoup4==4.12.3
behave==1.2.6
binaryornot==0.4.4
biopython==1.79
bleach==6.1.0
botocore==1.34.106
Bottleneck==1.3.2
build==1.2.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.2
cookiecutter==2.6.0
coverage==7.5.1
cycler==0.10.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
dnspython==2.6.1
docutils==0.20.1
dynaconf==3.2.5
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
execnet==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.25.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.4.0
gitdb==4.0.11
gitdb2==4.0.2
GitPython==3.0.6
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
graphlib_backport==1.1.0
graphql-core==3.2.3
greenlet==3.0.3
grimp==3.2
grpcio==1.39.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
identify==2.5.36
idna==3.2
imagesize==1.4.1
import-linter==2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipylab==1.0.0
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.0.1
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kedro==0.19.5
kedro-datasets==1.8.0
kedro-sphinx-theme==2024.4.0
kedro-telemetry==0.4.0
kedro-viz==7.1.0
keras==2.13.1
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282420641/work
lazy_loader==0.4
lazypredict==0.2.12
libclang==18.1.1
lightgbm==4.3.0
llvmlite==0.41.1
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1628003481308/work
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==10.2.0
multidict==6.0.5
munkres==1.1.4
mypy==1.10.0
mypy-extensions==1.0.0
myst-parser==2.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nodeenv==1.8.0
notebook==7.2.0
notebook_shim==0.2.4
numba==0.58.1
numexpr @ file:///tmp/build/80754af9/numexpr_1618856167419/work
numpy==1.22.4
nvidia-ml-py==11.460.79
oauthlib==3.1.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
omegaconf==2.3.0
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandas-stubs==2.0.3.230814
pandocfilters==1.5.1
parse==1.20.1
parse-type==0.6.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1625655817137/work
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.3.0
pre-commit==3.5.0
pre-commit-hooks==4.6.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==3.17.3
psutil==5.9.8
ptflops==0.6.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==1.10.15
Pygments==2.18.0
pynndescent==0.5.12
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyproject_hooks==1.1.0
pyspark==3.5.1
pytest==7.4.4
pytest-cov==3.0.0
pytest-forked==1.6.0
pytest-mock==1.13.0
pytest-xdist==2.2.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-slugify==8.0.4
pytoolconfig==1.3.1
pytorch-ignite==0.4.6
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.26.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rope==1.13.0
rpds-py==0.18.1
rsa==4.7.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.1.15
s3fs==2023.12.2
scikit-learn==0.24.2
scipy==1.7.1
screed==1.1.3
seaborn==0.13.2
secure==0.3.0
Send2Trash==1.8.3
shellingham==1.5.4
sip==4.19.13
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.1.2
sphinx-copybutton==0.5.2
sphinx-favicon==1.0.1
sphinx-last-updated-by-git==0.3.7
sphinx-notfound-page==1.0.1
sphinx-rtd-theme==2.0.0
sphinx_autodoc_typehints==1.20.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-youtube==1.4.1
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
strawberry-graphql==0.229.1
tenacity==8.3.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toposort==1.10
torch==1.9.0
torchaudio==0.9.0a0+33b2469
torchvision==0.10.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
truffleHog==2.2.1
truffleHogRegexes==0.0.7
typer==0.12.3
types-cachetools==5.3.0.7
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
types-toml==0.10.8.20240310
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
umap-learn==0.5.6
uri-template==1.3.0
urllib3==1.26.6
uvicorn==0.25.0
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.21.0
watchgod==0.8.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==2.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
xgboost==2.0.3
yacs==0.1.8
yarl==1.9.4
zipp==3.18.1
